Update django command for bulk data

+++ substrabac/substrapp/models/data.py
@@ -39,19 +39,23 @@ def my_handler(sender, instance, **kwargs):
     directory = path.join(getattr(settings, 'MEDIA_ROOT'), 'data/{0}'.format(instance.pk))
 
     # unzip file
-    zip_ref = zipfile.ZipFile(instance.path)
-    zip_ref.extractall(directory)
-    zip_ref.close()
+    if path.isfile(instance.path):
+        try:
+            zip_ref = zipfile.ZipFile(instance.path)
+            zip_ref.extractall(directory)
+            zip_ref.close()
+        except Exception as e:
+            logging.info(e)
 
-    # calculate new hash
-    sha256hash = dirhash(directory, 'sha256')
-    # rename directory to new hash if does not exist
-    new_directory = 'data/{0}'.format(sha256hash)
-    try:
-        rename(directory, path.join(getattr(settings, 'MEDIA_ROOT'), new_directory))
-    except Exception as e:
-        shutil.rmtree(directory)
-        logging.error(e, exc_info=True)
+        # calculate new hash
+        sha256hash = dirhash(directory, 'sha256')
+        # rename directory to new hash if does not exist
+        new_directory = 'data/{0}'.format(sha256hash)
+        try:
+            rename(directory, path.join(getattr(settings, 'MEDIA_ROOT'), new_directory))
+        except Exception as e:
+            shutil.rmtree(directory)
+            logging.error(e, exc_info=True)
 
-    instance.pkhash = sha256hash
-    instance.path = new_directory
+        instance.pkhash = sha256hash
+        instance.path = new_directory
